move token-related localstorage logic to toknes folder
