+++ package.json
@@ -78,6 +78,7 @@
     "react-dom": "^16.4.0",
     "react-final-form": "^4.1.0",
     "react-hot-loader": "^4.8.0",
+    "react-infinite-scroll-component": "^4.5.2",
     "react-redux": "^6.0.1",
     "react-router-dom": "^4.3.1",
     "recompose": "^0.30.0",
