+++ src/logic/tokens/store/actions/fetchTokens.js
@@ -8,7 +8,7 @@ import { getWeb3 } from '~/logic/wallets/getWeb3'
 import { type GlobalState } from '~/store/index'
 import { makeToken, type Token, type TokenProps } from '~/logic/tokens/store/model/token'
 import { ensureOnce } from '~/utils/singleton'
-import { getActiveTokenAddresses, getTokens } from '~/utils/localStorage/tokens'
+import { getActiveTokenAddresses, getTokens } from '~/logic/tokens/utils/activeTokensStorage'
 import { getSafeEthToken } from '~/utils/tokens'
 import { enhancedFetch } from '~/utils/fetch'
 import addTokens from './addTokens'
