+++ src/routes/safe/component/Balances/Tokens/index.jsx
@@ -4,6 +4,7 @@ import { connect } from 'react-redux'
 import { List } from 'immutable'
 import classNames from 'classnames/bind'
 import SearchBar from 'material-ui-search-bar'
+import InfiniteScroll from 'react-infinite-scroll-component'
 import { withStyles } from '@material-ui/core/styles'
 import MuiList from '@material-ui/core/List'
 import Img from '~/components/layout/Img'
