+++ src/test/tokens.dom.enabling.test.js
@@ -11,7 +11,7 @@ import { travelToTokens } from '~/test/builder/safe.dom.utils'
 import { sleep } from '~/utils/timer'
 import { buildMathPropsFrom } from '~/test/utils/buildReactRouterProps'
 import { tokenListSelector, activeTokensSelector } from '~/logic/tokens/store/selectors'
-import { getActiveTokenAddresses } from '~/utils/localStorage/tokens'
+import { getActiveTokenAddresses } from '~/logic/tokens/utils/activeTokensStorage'
 import { enableFirstToken, testToken } from '~/test/builder/tokens.dom.utils'
 import * as fetchTokensModule from '~/logic/tokens/store/actions/fetchTokens'
 import * as enhancedFetchModule from '~/utils/fetch'
